Create event listener for operators and equal buttons
Select all operators button and append event listener to all the
operators buttons to listen for click.

Once any operator button is clicked, the operator variable will be
assigned a value accoridngly and the value in display will be temporarily stored in the prevNum
variable.

The event listener for equal button will then call the operate function