Add conditional statement in the operator button event listener to handle over a pair of numbers
The conditional statement checks if the value of operator, the first
number and the second number are undefined. Their values should only be
undefined in the very beginning before any numbers or operators buttons
is clicked.

Once the first number, operator and second number have been assigned
values from clicking the buttons, the user has three options. The user
can either keep adding to the second number, or click on an operator, or
click on the equal button. If he clicks on the operator button with all
three variables with values already assigned, it will call the operate
function and display the results to be used for the next operation.