Fix bug when display changes when clicking on different operator buttons
When there is an interim or final display results shown on screen,
the display results would change from just clicking the operator buttons

This bug is fixed by introducing a new global variable prevButton to
track the class of last button clicked. If the last button clicked is
already an operator, then clicking onto operator again would just change
the operator global variable but not the variables for storing numbers.