Fix bug for pressing equal button before entering anything
NaN is displayed when equal button is pressed before anything is entered

The bug is fixed by adding a condition to the equal button event
listener. If prevNum is undefined, it means that it is either at new
start or clear stage so pressing the equal button will display the
presentNum which should be 0.