Fix the bug for adding decimal point after equal button is pressed
When equal button is pressed and result is displayed, pressing the
decimal point button would add decimal point to the display results.

The bug is fixed by resetting the global variables prevNum and operator,
while assigning presentNum to 0 so that it will store 0 as the prevNum
in case the user decides to press an operator button after putting in
any number after "0.".